Wrap The application contains one or more features that require access control that prevent IDOR.

+++ app/Http/Controllers/RecipeController.php
@@ -23,7 +23,11 @@ class RecipeController extends Controller
      */
     public function create()
     {
-        return view('recipes.create');
+        if (Auth::user()->hasRole('writer')) {
+            return view('recipes.create');
+        }
+
+        abort(403);
     }
 
     /**
@@ -31,14 +35,17 @@ class RecipeController extends Controller
      */
     public function store(Request $request)
     {
-        Recipe::create(array_merge(
-            $request->validate(Recipe::rules()),
-            ['creator_id' => Auth::user()->id],
-            ['photo' => $this->getPhoto($request)]
-        ));
+        if (Auth::user()->hasRole('writer')) {
+            Recipe::create(array_merge(
+                $request->validate(Recipe::rules()),
+                ['creator_id' => Auth::user()->id],
+                ['photo' => $this->getPhoto($request)]
+            ));
 
+            return redirect(route('recipes.index'));
+        }
 
-        return redirect(route('recipes.index'));
+        abort(403);
     }
 
     /**
@@ -54,7 +61,11 @@ class RecipeController extends Controller
      */
     public function edit(Recipe $recipe)
     {
-        return view('recipes.edit', compact('recipe'));
+        if (Auth::user()->id === $recipe->creator_id) {
+            return view('recipes.edit', compact('recipe'));
+        }
+
+        abort(403);
     }
 
     /**
@@ -62,15 +73,19 @@ class RecipeController extends Controller
      */
     public function update(Request $request, Recipe $recipe)
     {
-        $recipe->update(array_merge(
-            $request->validate(Recipe::rules()),
-        ));
+        if (Auth::user()->id === $recipe->creator_id) {
+            $recipe->update(array_merge(
+                $request->validate(Recipe::rules()),
+            ));
 
-        if ($request->has('photo')) {
-            $recipe->update(['photo' => $this->getPhoto($request)]);
+            if ($request->has('photo')) {
+                $recipe->update(['photo' => $this->getPhoto($request)]);
+            }
+
+            return redirect(route('recipes.show', $recipe))->with('status', 'The recipe has been successfully updated');
         }
 
-        return redirect(route('recipes.edit', $recipe))->with('status', 'The recipe has been successfully updated');
+        abort(403);
     }
 
     /**
@@ -78,9 +93,12 @@ class RecipeController extends Controller
      */
     public function destroy(Recipe $recipe)
     {
-        $recipe->delete();
+        if (Auth::user()->id === $recipe->creator_id || Auth::user()->hasRole('Admin')) {
+            $recipe->delete();
+            return redirect(route('recipes.index'))->with('status', 'The recipe has been successfully deleted');
+        }
 
-        return redirect(route('recipes.index'))->with('status', 'The recipe has been successfully deleted');
+        abort(403);
     }
 
     /**
